chore: move gateway to pkg/
